Gather files to deploy under _public.
This also eliminates the need to update app.yaml for each new document.

+++ build.sh
@@ -2,11 +2,21 @@
 # Convert markdown to HTML, prior to deploy.sh. Requires `marked':
 # https://www.npmjs.com/package/marked
 
-mkdir -p _build
+if [[ ! -f logo.svg ]]; then
+  echo >&2 "Unable to find Knola sources"
+  exit 1
+fi
+
+rm -rf _public
+mkdir -p _build _public _public/_resources
 
 for file in *.md; do
-  doc=_build/${file%.md}.html
+  base=${file%.md}
+  doc=_build/${base}.html
   cat header.html > $doc
   marked $file >> $doc
   cat footer.html >> $doc
+  ln $doc _public/$base 
 done
+
+ln logo.svg _public/_resources/logo.svg
